loading updates

+++ scgenome/cnplot.py
@@ -79,7 +79,7 @@ def plot_clustered_cell_cn_matrix(ax, cn_data, cn_field_name, cluster_field_name
     if not raw:
         cmap = get_cn_cmap(plot_data.values)
 
-    im = ax.imshow(plot_data.T, aspect='auto', cmap=cmap)
+    im = ax.imshow(plot_data.astype(float).T, aspect='auto', cmap=cmap)
 
     ax.set(xticks=chrom_mids)
     ax.set(xticklabels=utils.chrom_names)
